Revise ResourcesModel initialization
Refactor fetching semantics in the AbstractResourcesBackend

We used to have a reloadStarted and reloadFinished signals for keeping track
of the state of the backend. This changes makes it so we have a boolean
property called fetching that will be responsible of providing this
information.
This is better because it gives us further information about whether the
model is available for consumption. This is especially interesting for
writing unit tests and such tooling.

REVIEW: 111468