add keyboard shortcut to focus search box
Add keyboard shortcut to move focus to the search box, and give it focus at startup.

If the box is hidden, e.g. by changes preview, the focus is not changed. Otherwise you could change the content of the search box without seeing it.

(if accepted, please push it for me)

REVIEW: 126709
BUG: 299763
BUG: 312568
BUG: 325536