Fix muon not closing with window close button.
The variable m_canExit was not initialized and did prevent muon from exiting
when closed from the window manager. Remove it because it is never changed and
remove canExit() because it is in the wrong place anyway to prevent exiting
while acting on packages. It should be in backend.

Testing Done:
* open muon close by clicking the X button in the window decoration or with
* Alt+F4

Closing with File -> Exit or Ctrl+Q does not call MainWindow::queryClose() and
therefore did always work.

REVIEW: 127249
BUG: 358659