Connect to backend signals to catch cache reload.
When the cache is reloaded libqapt deletes Package objects and creates them
anew. This results in dangling pointers in muon models.

When "Preview Changes" is shown we have two models with lists of packages, and
both should be cleared and repopulated.

The cache reload can be triggered by pinning/unpinning (lock/unlock) a package.

REVIEW: 127107
BUG: 332530