Summary: emit currentPackageChanged just once
Summary:
Before, when a package was selected, `currentPackageChanged` was emitted
once from `currentChanged` and once from `selectionChanged` for the same
package.
Now it is emitted when the selection changes. It is still emitted for an
already-current package, but we need that to get the correct buttons
shown in `PackageWidget` when we deselect all packages except one from a
multi-selection.

Test Plan:
Add qDebug or breakpoint in `PackageWidget::packageActivated`.
Run and select packages in the list view.

Reviewers: sitter

Reviewed By: sitter

Differential Revision: https://phabricator.kde.org/D11152