Lazily refresh the details tabs
Summary:
Refresh only the current tab when the selected package is changed.
Before, every tab was refreshed. That included getting the list of
installed files and downloading the changelog, even if that information
was not presented to the user.

The UI has a small change in that before, when a tab had nothing to show
(e.g.  no alternative version available for the package) the tab itself
was hidden.  After the change, the tab is still visible but it is grayed
out and cannot be selected (i.e. inactive).

BUG: 390785

Reviewers: sitter

Reviewed By: sitter

Differential Revision: https://phabricator.kde.org/D11206